requirements.txt
streamlit
google-generativeai
python-dotenv
youtube-transcript-api
beautifulsoup4
requests
PyPDF2
pdfminer.six
langcodes
langchain
langchain-google-genai
langchain-community
langchain-huggingface
langchain-groq
sentence-transformers
faiss-cpu
selenium
webdriver-manager
pandas
openpyxl